requirements.txt
python-telegram-bot[webhooks]==20.7
python-dotenv==1.0.0
httpx==0.25.2
requests>=2.32.3
redis>=5.0.1
python-json-logger>=2.0.7
